Move GGTB to its own file
