+++ test/celeritas/GlobalTestBase.hh
@@ -10,7 +10,6 @@
 #include <iosfwd>
 #include <memory>
 #include <string>
-#include <gtest/gtest.h>
 
 #include "corecel/Assert.hh"
 #include "celeritas/geo/GeoParamsFwd.hh"
@@ -39,7 +38,7 @@ namespace celeritas_test
 /*!
  * Lazily construct core parameters, individually or together.
  *
- * \note Inherit from this class (or \c GeoTestBase below) using \c
+ * \note Inherit from this class (or \c GlobalGeoTestBase) using \c
  * virtual \c public so that tests can create mixins (see e.g. \c
  * SimpleStepperTest).
  */
@@ -138,47 +137,6 @@ class GlobalTestBase : public celeritas_test::Test
     SPConstRng         rng_;
     SPConstCore        core_;
     SPOutputManager    output_;
-};
-
-//---------------------------------------------------------------------------//
-/*!
- * Reuse geometry across individual tests.
- *
- * This is helpful for slow geometry construction or if the geometry has
- * trouble building/destroying multiple times per execution due to global
- * variable usage (VecGeom, Geant4).
- *
- * The "geometry basename" should be the filename without extension of a
- * geometry file inside `test/celeritas/data`.
- */
-class GlobalGeoTestBase : virtual public GlobalTestBase
-{
-  public:
-    // Overload with the base filename of the geometry
-    virtual const char* geometry_basename() const = 0;
-
-    // Construct a geometry that's persistent across tests
-    SPConstGeo build_geometry() final;
-
-    // Clear the lazy geometry
-    static void reset_geometry();
-
-  private:
-    //// LAZY GEOMETRY CONSTRUCTION AND CLEANUP FOR VECGEOM ////
-
-    struct LazyGeo
-    {
-        std::string basename{};
-        SPConstGeo  geo{};
-    };
-
-    static LazyGeo& lazy_geo();
-
-    class CleanupGeoEnvironment : public ::testing::Environment
-    {
-        void SetUp() override {}
-        void TearDown() override { GlobalGeoTestBase::reset_geometry(); }
-    };
 };
 
 //---------------------------------------------------------------------------//
