+++ test/celeritas/MockTestBase.hh
@@ -13,7 +13,7 @@
 #include "corecel/cont/Span.hh"
 #include "celeritas/Types.hh"
 
-#include "GlobalTestBase.hh"
+#include "GlobalGeoTestBase.hh"
 
 namespace celeritas
 {
