+++ test/celeritas/SimpleTestBase.hh
@@ -9,7 +9,7 @@
 
 #include "corecel/Types.hh"
 
-#include "GlobalTestBase.hh"
+#include "GlobalGeoTestBase.hh"
 
 namespace celeritas_test
 {
